oms into common comp
