+++ services/ui-src/src/measures/2021/CommonQuestions/OtherPerformanceMeasure/index.tsx
@@ -81,7 +81,7 @@ export const OtherPerformanceMeasure = ({
           }}
           onClick={() => {
             showRates.push({
-              description: [""],
+              description: "",
               rate: [
                 {
                   denominator: "",
