+++ services/ui-src/src/measures/2021/CommonQuestions/index.tsx
@@ -6,3 +6,4 @@ export * from "./OtherPerformanceMeasure";
 export * from "./Reporting";
 export * from "./StatusOfData";
 export * from "./WhyAreYouNotReporting";
+export * from "./OptionalMeasureStrat";
