+++ services/ui-src/src/measures/2021/FUAAD/validation/types.ts
@@ -19,7 +19,7 @@ export namespace Measure {
   }
 
   interface OtherRatesFields {
-    description: string[];
+    description: string;
     rate: RateFields[];
   }
   export interface Form
