+++ services/ui-src/src/measures/2021/FUMAD/validation/types.ts
@@ -21,7 +21,7 @@ export namespace Measure {
   }
 
   interface OtherRatesFields {
-    description: string[];
+    description: string;
     rate: RateFields[];
   }
