+++ services/ui-src/src/measures/2021/IETAD/questions/index.tsx
@@ -2,5 +2,3 @@ export * from "./DataSource";
 export * from "./DeviationFromMeasureSpec";
 export * from "./MeasurementSpecification";
 export * from "./PerformanceMeasure";
-export * from "./OptionalMeasureStratification";
-export * from "./OptionalMeasureStrat";
