+++ services/ui-src/src/measures/2021/IETAD/validation/types.ts
@@ -7,76 +7,6 @@ export namespace Measure {
     setValidationFunctions?: React.Dispatch<React.SetStateAction<any>>;
   }
 
-  export interface RateFields {
-    numerator?: string;
-    denominator?: string;
-    rate?: string;
-  }
-
-  interface AggregateRate {
-    subRate?: RateFields[];
-    total?: RateFields[];
-  }
-
-  export interface OtherRatesFields {
-    description?: string;
-    rate?: RateFields[];
-  }
-
-  interface OmsRateFields {
-    options?: string[];
-    rates?: {
-      /** rate label will be some combination of ageRange_perfDesc or opmFieldLabel */
-      [rateLabel: string]: RateFields[];
-    };
-    total?: RateFields[];
-  }
-
-  interface LowLevelOmsNode {
-    // if just ndr sets
-    ageRangeRates?: OmsRateFields;
-
-    // for additional subCats/add anothers
-    subCatOptions?: string[];
-    subCategories?: {
-      description?: string;
-      ageRangeRates?: OmsRateFields;
-    }[];
-  }
-
-  interface MidLevelOMSNode extends LowLevelOmsNode {
-    // if sub-options
-    aggregate?: string;
-    options?: string[];
-    selections?: {
-      [option: string]: LowLevelOmsNode;
-    };
-  }
-
-  interface TopLevelOmsNode {
-    // top level child, ex: Race, Sex, Ethnicity
-    options?: string[]; // checkbox
-    additionalCategories?: string[]; // add another section
-    selections?: {
-      [option: string]: MidLevelOMSNode;
-    };
-    additionalSelections?: AddtnlOmsNode[];
-
-    // catch case for ACA
-    ageRangeRates?: OmsRateFields;
-  }
-
-  interface AddtnlOmsNode extends LowLevelOmsNode {
-    description?: string;
-  }
-
-  interface DeviationFields {
-    options: string[];
-    denominator: string;
-    numerator: string;
-    other: string;
-  }
-
   export interface Form
     extends Types.DefinitionOfPopulation,
       Types.DateRange,
@@ -106,7 +36,7 @@ export namespace Measure {
 
     //Other Performance Measure
     "OtherPerformanceMeasure-Explanation": string;
-    "OtherPerformanceMeasure-Rates": OtherRatesFields[];
+    "OtherPerformanceMeasure-Rates": Types.OtherRatesFields[];
     "OtherPerformanceMeasure-Notes": string;
     "OtherPerformanceMeasure-Rates-TextInput": string;
 
@@ -114,106 +44,38 @@ export namespace Measure {
     DidCalculationsDeviate: string;
     DeviationOptions: string[];
     "DeviationOptions-InitAlcohol-AgeRange": string[];
-    "DeviationFields-InitAlcohol": DeviationFields;
+    "DeviationFields-InitAlcohol": Types.DeviationFields;
 
     "DeviationOptions-EngageAlcohol-AgeRange": string[];
-    "DeviationFields-EngageAlcohol": DeviationFields;
+    "DeviationFields-EngageAlcohol": Types.DeviationFields;
 
     "DeviationOptions-InitOpioid-AgeRange": string[];
-    "DeviationFields-InitOpioid": DeviationFields;
+    "DeviationFields-InitOpioid": Types.DeviationFields;
 
     "DeviationOptions-EngageOpioid-AgeRange": string[];
-    "DeviationFields-EngageOpioid": DeviationFields;
+    "DeviationFields-EngageOpioid": Types.DeviationFields;
 
     "DeviationOptions-InitOther-AgeRange": string[];
-    "DeviationFields-InitOther": DeviationFields;
+    "DeviationFields-InitOther": Types.DeviationFields;
 
     "DeviationOptions-EngageOther-AgeRange": string[];
-    "DeviationFields-EngageOther": DeviationFields;
+    "DeviationFields-EngageOther": Types.DeviationFields;
 
     "DeviationOptions-InitTotal-AgeRange": string[];
-    "DeviationFields-InitTotal": DeviationFields;
+    "DeviationFields-InitTotal": Types.DeviationFields;
 
     "DeviationOptions-EngageTotal-AgeRange": string[];
-    "DeviationFields-EngageTotal": DeviationFields;
+    "DeviationFields-EngageTotal": Types.DeviationFields;
 
     //PerformanceMeasure
     "PerformanceMeasure-Explanation": string;
-    "PerformanceMeasure-AgeRates-Initiation-Alcohol": RateFields[];
-    "PerformanceMeasure-AgeRates-Engagement-Alcohol": RateFields[];
-    "PerformanceMeasure-AgeRates-Initiation-Opioid": RateFields[];
-    "PerformanceMeasure-AgeRates-Engagement-Opioid": RateFields[];
-    "PerformanceMeasure-AgeRates-Initiation-Other": RateFields[];
-    "PerformanceMeasure-AgeRates-Engagement-Other": RateFields[];
-    "PerformanceMeasure-AgeRates-Initiation-Total": RateFields[];
-    "PerformanceMeasure-AgeRates-Engagement-Total": RateFields[];
-
-    //OptionalMeasureStratification
-    CategoriesReported: string[];
-
-    AddtnlEthnicity: string[];
-    AddtnlEthnicityRates: AggregateRate[];
-
-    AddtnlNonHispanicRace: string[];
-    AddtnlNonHispanicRaceRates: AggregateRate[];
-    AddtnlNonHispanicRaceSubCatTitle: { titles: string[] }[];
-    AddtnlNonHispanicRaceSubCatOptions: string[][];
-    AddtnlNonHispanicRaceSubCatRates: { rates: AggregateRate[] }[];
-
-    AddtnlNonHispanicSubCat: string[];
-    AddtnlNonHispanicSubCatRates: AggregateRate[];
-
-    NonHispanicRacialCategories: string[];
-    "NHRC-WhiteRates": AggregateRate;
-    "NHRC-BlackOrAfricanAmericanRates": AggregateRate;
-    "NHRC-AmericanIndianOrAlaskaNativeRates": AggregateRate;
-    "NHRC-AggregateAsianRates": AggregateRate;
-    "NHRC-IndependentAsianRates": AggregateRate[];
-    "NHRC-AggregateHawaiianOrPacificIslanderRates": AggregateRate;
-    "NHRC-IndependentHawaiianOrPacificIslanderRates": AggregateRate[];
-
-    EthnicityCategories: string[];
-    EthnicitySubCategories: string[];
-    NonHispanicEthnicityRates: AggregateRate;
-    HispanicIndependentReporting: string;
-    HispanicEthnicityAggregateRate: AggregateRate;
-    IndependentHispanicOptions: string[];
-    IndependentHispanicRates: AggregateRate[];
-
-    AsianIndependentReporting: string;
-    IndependentAsianOptions: string[];
-    NativeHawaiianIndependentReporting: string;
-    IndependentNativeHawaiianOptions: string[];
-
-    SexOptions: string[];
-    MaleSexRates: AggregateRate;
-    FemaleSexRates: AggregateRate;
-
-    PrimaryLanguageOptions: string[];
-    AddtnlPrimaryLanguage: string[];
-    AddtnlPrimaryLanguageRates: AggregateRate[];
-    EnglishLanguageRate: AggregateRate;
-    SpanishLanguageRate: AggregateRate;
-
-    DisabilityStatusOptions: string[];
-    DisabilitySSIRate: AggregateRate;
-    DisabilityNonSSIRate: AggregateRate;
-    AddtnlDisabilityStatusDesc: string;
-    AddtnlDisabilityRate: AggregateRate;
-
-    GeographyOptions: string[];
-    UrbanGeographyRate: AggregateRate;
-    RuralGeographyRate: AggregateRate;
-    AddtnlGeographyDesc: string;
-    AddtnlGeographyRate: AggregateRate;
-
-    ACAGroupRate: AggregateRate;
-
-    OptionalMeasureStratification: {
-      options: string[]; //checkbox
-      selections: {
-        [option: string]: TopLevelOmsNode;
-      };
-    };
+    "PerformanceMeasure-AgeRates-Initiation-Alcohol": Types.RateFields[];
+    "PerformanceMeasure-AgeRates-Engagement-Alcohol": Types.RateFields[];
+    "PerformanceMeasure-AgeRates-Initiation-Opioid": Types.RateFields[];
+    "PerformanceMeasure-AgeRates-Engagement-Opioid": Types.RateFields[];
+    "PerformanceMeasure-AgeRates-Initiation-Other": Types.RateFields[];
+    "PerformanceMeasure-AgeRates-Engagement-Other": Types.RateFields[];
+    "PerformanceMeasure-AgeRates-Initiation-Total": Types.RateFields[];
+    "PerformanceMeasure-AgeRates-Engagement-Total": Types.RateFields[];
   }
 }
