+++ services/ui-src/src/measures/2021/LBWCH/index.tsx
@@ -1,7 +1,11 @@
-import { Measure } from "measures/types";
 import * as QMR from "components";
 
-export const LBWCH = ({ name, year }: Measure.Props) => {
+interface Props {
+  name: string;
+  year: string;
+}
+
+export const LBWCH = ({ name, year }: Props) => {
   return (
     <QMR.AutocompletedMeasureTemplate
       year={year}
