+++ services/ui-src/src/measures/2021/PQI05AD/validation/types.ts
@@ -22,7 +22,7 @@ export namespace Measure {
   }
 
   interface OtherRatesFields {
-    description: string[];
+    description: string;
     rate: RateFields[];
   }
