+++ services/ui-src/src/measures/2021/PQI08AD/validation/types.ts
@@ -21,7 +21,7 @@ export namespace Measure {
   }
 
   interface OtherRatesFields {
-    description: string[];
+    description: string;
     rate: RateFields[];
   }
